+++ PSFramework/internal/tepp/tepp-assignment.ps1
@@ -1,5 +1,4 @@
-﻿
-# Configuration
+﻿#region Configuration
 Register-PSFTeppArgumentCompleter -Command Export-PSFConfig -Parameter FullName -Name 'PSFramework-config-fullname'
 Register-PSFTeppArgumentCompleter -Command Export-PSFConfig -Parameter Module -Name 'PSFramework-config-module'
 Register-PSFTeppArgumentCompleter -Command Export-PSFConfig -Parameter Name -Name 'PSFramework-config-name'
@@ -21,3 +20,8 @@ Register-PSFTeppArgumentCompleter -Command Get-PSFConfigValue -Parameter FullNam
 
 Register-PSFTeppArgumentCompleter -Command Unregister-PSFConfig -Parameter FullName -Name 'PSFramework-Unregister-PSFConfig-FullName'
 Register-PSFTeppArgumentCompleter -Command Unregister-PSFConfig -Parameter Module -Name 'PSFramework-Unregister-PSFConfig-Module'
+#endregion Configuration
+
+#region Tab Completion
+Register-PSFTeppArgumentCompleter -Command Set-PSFTeppResult -Parameter TabCompletion -Name 'PSFramework-tepp-scriptblockname'
+#endregion Tab Completion